+++ devel/projects/spa/src/hr/restart/os/repLikKum.java
@@ -98,7 +98,7 @@ public class repLikKum implements sg.com.elixir.reportwriter.datasource.IDataPro
 
   public double getSaldo()
   {
-    return ds.getDouble("SALDO");
+    return ds.getBigDecimal("SALDO").doubleValue();
   }
 
   public String getNazOrg()
@@ -117,12 +117,12 @@ public class repLikKum implements sg.com.elixir.reportwriter.datasource.IDataPro
 
   public double getIspravak()
   {
-    return ds.getDouble("ISPRAVAK");
+    return ds.getBigDecimal("ISPRAVAK").doubleValue();
   }
 
   public double getOsnovica()
   {
-    return ds.getDouble("OSNOVICA");
+    return ds.getBigDecimal("OSNOVICA").doubleValue();
   }
